+++ src/udpqueue.rs
@@ -141,7 +141,7 @@ impl UdpQueue {
                             // Not exactly the correct error, but has similar meaning
                             Err(ErrorKind::WriteZero.into())
                         }
-                        Ok(_) => Ok(()),
+                        Ok(_) => break,
                     };
                 }
             }